We do NOT put .gitignore in our repositories!! X'D
It's a list of things that are specific to your system and not to the
project!  So it would be different on different peoples' systems
obviously!

I yearn for the day when Eclipse allows a "pub/" subfolder in a project
to be the repository instead of the project folder itself, then I won't
need to worry about .gitignores at all! :'3

(eg, if one person uses Eclipse and needs .project / .classpath and
another person uses something else like NetBeans or Xcode and they put
different files there! or who knows what!  let people be different!
don't make the list of non-repo files that can be different for
different users be a part of the repo that's the same for every user!
kdjsdlfdksjfldk)